+++ GameServerConsole/Settings/GameInfo.json.template
@@ -1,62 +1,63 @@
 {
-    "players": [
-        {
-            "playerId": 1,
-            "blowfishKey": "17BLOhi6KZsTtldTsizvHg==",
-            "rank": "DIAMOND",
-            "name": "Test",
-            "champion": "Ezreal",
-            "team": "BLUE",
-            "skin": 0,
-            "summoner1": "SummonerHeal",
-            "summoner2": "SummonerFlash",
-            "ribbon": 2,
-            "icon": 0,
-            "runes": {
-                //DO NOT CHANGE THESE IF YOU DONT KNOW WHAT YOU ARE DOING.
-                "1": 5245,
-                "2": 5245,
-                "3": 5245,
-                "4": 5245,
-                "5": 5245,
-                "6": 5245,
-                "7": 5245,
-                "8": 5245,
-                "9": 5245,
-                "10": 5317,
-                "11": 5317,
-                "12": 5317,
-                "13": 5317,
-                "14": 5317,
-                "15": 5317,
-                "16": 5317,
-                "17": 5317,
-                "18": 5317,
-                "19": 5289,
-                "20": 5289,
-                "21": 5289,
-                "22": 5289,
-                "23": 5289,
-                "24": 5289,
-                "25": 5289,
-                "26": 5289,
-                "27": 5289,
-                "28": 5335,
-                "29": 5335,
-                "30": 5335
-            }
-        }
-    ],
-    "game": {
-        "map": 1,
-        "dataPackage": "LeagueSandbox-Scripts"
-    },
-    "gameInfo": {
-        "MANACOSTS_ENABLED": false,
-        "COOLDOWNS_ENABLED": false,
-        "CHEATS_ENABLED": true,
-        "MINION_SPAWNS_ENABLED": false,
-        "CONTENT_PATH": "../../../../../Content",
-        "IS_DAMAGE_TEXT_GLOBAL": false
+  "players": [
+    {
+      "playerId": 1,
+      "blowfishKey": "17BLOhi6KZsTtldTsizvHg==",
+      "rank": "DIAMOND",
+      "name": "Test",
+      "champion": "Ezreal",
+      "team": "BLUE",
+      "skin": 0,
+      "summoner1": "SummonerHeal",
+      "summoner2": "SummonerFlash",
+      "ribbon": 2,
+      "icon": 0,
+      "runes": {
+        //DO NOT CHANGE THESE IF YOU DONT KNOW WHAT YOU ARE DOING.
+        "1": 5245,
+        "2": 5245,
+        "3": 5245,
+        "4": 5245,
+        "5": 5245,
+        "6": 5245,
+        "7": 5245,
+        "8": 5245,
+        "9": 5245,
+        "10": 5317,
+        "11": 5317,
+        "12": 5317,
+        "13": 5317,
+        "14": 5317,
+        "15": 5317,
+        "16": 5317,
+        "17": 5317,
+        "18": 5317,
+        "19": 5289,
+        "20": 5289,
+        "21": 5289,
+        "22": 5289,
+        "23": 5289,
+        "24": 5289,
+        "25": 5289,
+        "26": 5289,
+        "27": 5289,
+        "28": 5335,
+        "29": 5335,
+        "30": 5335
+      }
     }
+  ],
+  "game": {
+    "map": 1,
+    "GameMode": "CLASSIC",
+    "dataPackage": "LeagueSandbox-Scripts"
+  },
+  "gameInfo": {
+    "MANACOSTS_ENABLED": false,
+    "COOLDOWNS_ENABLED": false,
+    "CHEATS_ENABLED": true,
+    "MINION_SPAWNS_ENABLED": false,
+    "CONTENT_PATH": "../../../../../Content",
+    "IS_DAMAGE_TEXT_GLOBAL": false
+  }
 }
