+++ GameServerCore/Domain/IGlobalData.cs
@@ -8,6 +8,7 @@
         float PercentAttackSpeedModMinimum { get; set; }
         float AttackMaxDelay { get; set; }
         float CooldownMinimum { get; set; }
+        float PercentCooldownModMinimun { get; }
         float PercentRespawnTimeModMinimum { get; set; }
         float PercentGoldLostOnDeathModMinimum { get; set; }
         float PercentEXPBonusMinimum { get; set; }
