+++ GameServerLib/Content/CharData.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Text.RegularExpressions;
 using GameServerCore.Domain;
 using GameServerCore.Enums;
+using GameServerLib.GameObjects.GlobalData;
 using LeagueSandbox.GameServer.Logging;
 using log4net;
 using Newtonsoft.Json;
@@ -11,20 +12,6 @@ using Newtonsoft.Json;
 namespace LeagueSandbox.GameServer.Content
 {
     // TODO: Add Constants.var files to each Map's Content folder and assign values by reading them, currently this Data is only for Map1 as a placeholder.
-    public class GlobalData : IGlobalData
-    {
-        public float AttackDelay { get; set; } = 1.600f;
-        public float AttackDelayCastPercent { get; set; } = 0.300f;
-        public float AttackMinDelay { get; set; } = 0.400f;
-        public float PercentAttackSpeedModMinimum { get; set; } = -0.950f;
-        public float AttackMaxDelay { get; set; } = 5.000f;
-        public float CooldownMinimum { get; set; } = 0.000f;
-        public float PercentRespawnTimeModMinimum { get; set; } = -0.950f;
-        public float PercentGoldLostOnDeathModMinimum { get; set; } = -0.950f;
-        public float PercentEXPBonusMinimum { get; set; } = -1.000f;
-        public float PercentEXPBonusMaximum { get; set; } = 5.000f;
-    }
-
     public class PassiveData : IPassiveData
     {
         public string PassiveAbilityName { get; set; } = "";
