+++ GameServerLib/GameObjects/Spell/Spell.cs
@@ -185,7 +185,7 @@ namespace LeagueSandbox.GameServer.GameObjects.Spell
 
             CastInfo.AttackSpeedModifier = stats.AttackSpeedMultiplier.Total;
 
-            if (_game.Config.ManaCostsEnabled)
+            if (_game.Config.GameFeatures.ManaCostsEnabled)
             {
                 stats.CurrentMana -= SpellData.ManaCost[CastInfo.SpellLevel] * (1 - stats.SpellCostReduction);
             }
@@ -434,7 +434,7 @@ namespace LeagueSandbox.GameServer.GameObjects.Spell
                     return false;
                 }
 
-                if (_game.Config.ManaCostsEnabled)
+                if (_game.Config.GameFeatures.ManaCostsEnabled)
                 {
                     stats.CurrentMana -= SpellData.ManaCost[CastInfo.SpellLevel] * (1 - stats.SpellCostReduction);
                 }
@@ -966,7 +966,7 @@ namespace LeagueSandbox.GameServer.GameObjects.Spell
 
         public float GetCooldown()
         {
-            return _game.Config.CooldownsEnabled ? SpellData.Cooldown[CastInfo.SpellLevel] * (1 - CastInfo.Owner.Stats.CooldownReduction.Total) : 0;
+            return _game.Config.GameFeatures.CooldownsEnabled ? SpellData.Cooldown[CastInfo.SpellLevel] * (1 - CastInfo.Owner.Stats.CooldownReduction.Total) : 0;
         }
 
         public float GetCurrentCastRange()
