+++ GameServerLib/Scripting/CSharp/EmptyMapScript.cs
@@ -5,11 +5,14 @@ using GameServerCore.Domain;
 using GameServerCore.Domain.GameObjects;
 using GameServerCore.Enums;
 using GameServerCore.Maps;
+using GameServerLib.GameObjects.GlobalData;
 
 namespace MapScripts
 {
     public class EmptyMapScript : IMapScript
     {
+        public virtual IGlobalData GlobalData { get; set; } = new GlobalData();
+        public virtual IGameFeatures GameFeatures { get; set; }
         public bool EnableBuildingProtection { get; set; } = false;
 
         //General Map variable
